+++ CHANGELOG.md
@@ -18,7 +18,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-> nothing new
+### Added
+- `galapy.analysis.plot.corner_derived` function for plotting the marginal 2d posterior
+  of derived parameters. The API is completely equivalent to the one
+  used for the free parameters
+
+### Fixed
+- `MANIFEST.in` now includes the *.tpp extensions used for templated headers of the
+  C++ library sector
+- support for more out-of-the-box architectures is now made available (WSL compliant
+  and intel-based MacOS architectures
 
 ## [0.5.7] - 2026-06-11